BeachedWhale 8.0 removes Windows 10 Bloatware
BeachedWhale 8.0 is the first public release of my personal project.

Windows 10 comes with a ton of bloatware applications that no one wants or needs, especially in a business environment. BeachedWhale is designed to remove it easily.

The GUI is written in Python using the appJar repo that's bundled with this version. The removal off apps and programs is handled with PowerShell scripts that are customizable if you understand the syntax.

Let me know what you think!